runmod

Struct RunMod

pub struct RunMod { /* private fields */ }

Implementations

impl RunMod

pub fn new(val: RunVar) -> Self

Makes a new RunMod

This method will never panic or fail

Examples

use runmod::{RunMod, RunVar};

let value = RunMod::new(RunVar::I32(10));

pub fn get_i8(&mut self) -> Option<i8>

Reads the source file andf gets a new value. I recommend that you disable autosave on your text editor/IDE (especily if it is high frequency), as this can cause unexpected panics.

This will update every time you call it, witch means it will open the file, read it, parse the line, then close it.

pub fn get_i16(&mut self) -> Option<i16>

Reads the source file andf gets a new value. I recommend that you disable autosave on your text editor/IDE (especily if it is high frequency), as this can cause unexpected panics.

This will update every time you call it, witch means it will open the file, read it, parse the line, then close it.

pub fn get_i32(&mut self) -> Option<i32>

Reads the source file andf gets a new value. I recommend that you disable autosave on your text editor/IDE (especily if it is high frequency), as this can cause unexpected panics.

This will update every time you call it, witch means it will open the file, read it, parse the line, then close it.

pub fn get_i64(&mut self) -> Option<i64>

Reads the source file andf gets a new value. I recommend that you disable autosave on your text editor/IDE (especily if it is high frequency), as this can cause unexpected panics.

This will update every time you call it, witch means it will open the file, read it, parse the line, then close it.

pub fn get_i128(&mut self) -> Option<i128>

Reads the source file andf gets a new value. I recommend that you disable autosave on your text editor/IDE (especily if it is high frequency), as this can cause unexpected panics.

This will update every time you call it, witch means it will open the file, read it, parse the line, then close it.

pub fn get_isize(&mut self) -> Option<isize>

Reads the source file andf gets a new value. I recommend that you disable autosave on your text editor/IDE (especily if it is high frequency), as this can cause unexpected panics.

This will update every time you call it, witch means it will open the file, read it, parse the line, then close it.

pub fn get_u8(&mut self) -> Option<u8>

Reads the source file andf gets a new value. I recommend that you disable autosave on your text editor/IDE (especily if it is high frequency), as this can cause unexpected panics.

This will update every time you call it, witch means it will open the file, read it, parse the line, then close it.

pub fn get_u16(&mut self) -> Option<u16>

Reads the source file andf gets a new value. I recommend that you disable autosave on your text editor/IDE (especily if it is high frequency), as this can cause unexpected panics.

This will update every time you call it, witch means it will open the file, read it, parse the line, then close it.

pub fn get_u32(&mut self) -> Option<u32>

Reads the source file andf gets a new value. I recommend that you disable autosave on your text editor/IDE (especily if it is high frequency), as this can cause unexpected panics.

This will update every time you call it, witch means it will open the file, read it, parse the line, then close it.

pub fn get_u64(&mut self) -> Option<u64>

Reads the source file andf gets a new value. I recommend that you disable autosave on your text editor/IDE (especily if it is high frequency), as this can cause unexpected panics.

This will update every time you call it, witch means it will open the file, read it, parse the line, then close it.

pub fn get_u128(&mut self) -> Option<u128>

Reads the source file andf gets a new value. I recommend that you disable autosave on your text editor/IDE (especily if it is high frequency), as this can cause unexpected panics.

This will update every time you call it, witch means it will open the file, read it, parse the line, then close it.

pub fn get_usize(&mut self) -> Option<usize>

Reads the source file andf gets a new value. I recommend that you disable autosave on your text editor/IDE (especily if it is high frequency), as this can cause unexpected panics.

This will update every time you call it, witch means it will open the file, read it, parse the line, then close it.

pub fn get_f32(&mut self) -> Option<f32>

Reads the source file andf gets a new value. I recommend that you disable autosave on your text editor/IDE (especily if it is high frequency), as this can cause unexpected panics.

This will update every time you call it, witch means it will open the file, read it, parse the line, then close it.

pub fn get_f64(&mut self) -> Option<f64>

Reads the source file andf gets a new value. I recommend that you disable autosave on your text editor/IDE (especily if it is high frequency), as this can cause unexpected panics.

This will update every time you call it, witch means it will open the file, read it, parse the line, then close it.